pyproject.toml
[project]
name = "agenticgemini"
version = "0.1.0"
description = "An autonomous agent framework powered by Google Gemini models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
